Reduce the scope of helper class
