+++ test/System/FileInputStream.cpp
@@ -1,9 +1,61 @@
 #include <SFML/System/FileInputStream.hpp>
-#include "SystemUtil.hpp"
+#include <cassert>
+#include <fstream>
+#include <sstream>
 #include <string_view>
 #include <utility>
 
 #include <doctest.h>
+    
+static std::string getTemporaryFilePath()
+{
+    static int counter = 0;
+
+    std::ostringstream oss;
+    oss << "sfmltemp" << counter << ".tmp";
+    ++counter;
+
+    std::filesystem::path result;
+    result /= std::filesystem::temp_directory_path();
+    result /= oss.str();
+
+    return result.string();
+}
+
+class TemporaryFile
+{
+private:
+    std::string m_path;
+
+public:
+    // Create a temporary file with a randomly generated path, containing 'contents'.
+    TemporaryFile(const std::string& contents)
+        : m_path(getTemporaryFilePath())
+    {
+        std::ofstream ofs(m_path);
+        assert(ofs);
+
+        ofs << contents;
+        assert(ofs);
+    }
+
+    // Close and delete the generated file.
+    ~TemporaryFile()
+    {
+        [[maybe_unused]] const bool removed = std::filesystem::remove(m_path);
+        assert(removed);
+    }
+
+    // Prevent copies.
+    TemporaryFile(const TemporaryFile&) = delete;
+    TemporaryFile& operator=(const TemporaryFile&) = delete;
+
+    // Return the randomly generated path.
+    const std::string& getPath() const
+    {
+        return m_path;
+    }
+};
 
 TEST_CASE("sf::FileInputStream class - [system]")
 {
@@ -20,7 +72,7 @@ TEST_CASE("sf::FileInputStream class - [system]")
     {
         const std::string fileContents = "hello world";
 
-        sf::Testing::TemporaryFile tmpFile(fileContents);
+        TemporaryFile tmpFile(fileContents);
         sf::FileInputStream fis;
 
         REQUIRE(fis.open(tmpFile.getPath()));
